requirements.txt
beautifulsoup4>=4.13.4
discord-py>=2.5.2
requests>=2.32.4
trafilatura>=2.0.0
flask
